+++ XYZHubConnector/modules/controller/__init__.py
@@ -10,7 +10,7 @@
 
 
 from ..common.error import make_exception_obj, parse_exception_obj
-from ..common.signal import (BasicSignal, LoggingSignal, make_qt_args, output_to_qt_args,
+from ..common.signal import (BasicSignal, make_qt_args, output_to_qt_args,
                       parse_qt_args, make_fun_args)
                       
 from .thread_safe.async_fun import AsyncFun, WorkerFun, NetworkFun