+++ XYZHubConnector/modules/layer/parser.py
@@ -25,6 +25,21 @@ print_qgis = make_print_qgis("parser")
 QGS_XYZ_ID = "xyz_id"
 XYZ_ID = "id"
 
+QGS_ID = "fid"
+
+
+def unique_field_name(name, i):
+    if name.startswith("@"):
+        return name
+    return name + ".%s"%i 
+def normal_field_name(name):
+    if name.startswith("@"):
+        return name
+    parts = name.split(".")
+    if len(parts) > 1 and parts[-1].isdigit():
+        return ".".join(parts[0:-1])
+    return name
+
 LIMIT = int(1e7) # Amazon API limit: 10485760
 def estimate_chunk_size(byt):
     chunk_size = LIMIT // len(byt) # round down
@@ -41,7 +56,7 @@ def estimate_upload_chunk_single(lst):
     return chunk_size
 def make_lst_feature_collection(features):
     if len(features) == 0: 
-        return None
+        return list()
     chunk_size = estimate_upload_chunk_single(features)
     def _iter_collection():
         i0, i1= 0,0
@@ -76,15 +91,23 @@ def feature_collection(features):
 
 def feature_to_xyz_json(feature, vlayer, is_new=False):
     def _xyz_props(props):
-        # for all key start with @: str to dict
+        # for all key start with @ (internal): str to dict (disabled)
         # k = "@ns:com:here:xyz"
-        for k in props.keys():
-            if not k.startswith("@"): continue
-            v = props[k]
-            if isinstance(v,str):
-                try: props[k] = json.loads(v)
-                except json.JSONDecodeError: pass # naively handle error
-        return props
+        new_props = dict()
+        for t in props.keys():
+            # drop @ field, for consistency
+            if t.startswith("@ns:com:here:xyz"): continue
+            k = normal_field_name(t)
+            new_props[k] = props[t]
+
+            # # disabled
+            # if not k.startswith("@"): continue
+            # v = new_props[k]
+            # if isinstance(v,str): # no need to handle json str in props
+            #     # print_qgis(json.dumps(dict(v=v), ensure_ascii=False))
+            #     try: new_props[k] = json.loads(v)
+            #     except json.JSONDecodeError: pass # naively handle error
+        return new_props
     def _single_feature(feat, transformer):
         # existing feature json
         obj = {
@@ -92,10 +115,15 @@ def feature_to_xyz_json(feature, vlayer, is_new=False):
         }
         json_str = QgsJsonUtils.exportAttributes(feat)
         props = json.loads(json_str)
-        props = _xyz_props(props)
-        if QGS_XYZ_ID in props:
-            v = props.pop(QGS_XYZ_ID)
+        props.pop(QGS_ID,"")
+        v = props.pop(QGS_XYZ_ID,"")
+        if len(v) > 0:
+            if v in exist_feat_id:
+                return None
+            exist_feat_id.add(v)
             if not is_new: obj[XYZ_ID] = v
+                
+        props = _xyz_props(props)
         obj["properties"] = props
 
         geom = feat.geometry()
@@ -120,10 +148,11 @@ def feature_to_xyz_json(feature, vlayer, is_new=False):
     crs_src = vlayer.crs()
     crs_dst = QgsCoordinateReferenceSystem('EPSG:4326')
     transformer = QgsCoordinateTransform(crs_src, crs_dst, QgsProject.instance())
-    return list( 
+    exist_feat_id = set()
+    return list( filter(None,(
         _single_feature(ft, transformer) 
         for ft in feature if ft.hasGeometry() # FIX: XYZHub doesnt like empty geom
-    )
+    )))
 
 # https://github.com/qgis/QGIS/blob/f3e9aaf79a9282b28a605abd0dadaab9951050c8/python/plugins/processing/algs/qgis/ui/FieldsMappingPanel.py
 valid_fieldTypes = dict([
@@ -183,6 +212,8 @@ def xyz_json_to_feature(txt, map_fields=dict()):
         feat=QgsFeature()
         
         names = fields.names()
+        if QGS_ID in names: names.remove(QGS_ID)
+        names_normal = list(map(normal_field_name, names))
         qattrs = list()
 
         # handle xyz id
@@ -210,8 +241,12 @@ def xyz_json_to_feature(txt, map_fields=dict()):
                 if not val.type() in valid_qvariant:
                     print_qgis("Invalid type", k, val.typeName())
                     continue
-                if k not in names:
+                if k not in names_normal:
+                    k = unique_field_name(k, len(fields))
                     fields.append( make_field(k, val))
+                else:
+                    idx = names_normal.index(k)
+                    k = names[idx]
                 qattrs.append([k,val])
 
             feat.setFields(fields)